refactor: move address-validation-test from `no-meteor` to `node-app`
